refactor: 더미 데이터 수정
- 인터페이스를 만들어 나중에 유지보수를 용이하게 하였다.
- domain으로 이동
