+++ app/src/main/java/woowacourse/movie/domain/data/MovieData.kt
@@ -1,0 +1,7 @@
+package woowacourse.movie.domain.data
+
+import woowacourse.movie.domain.movieinfo.Movie
+
+interface MovieData {
+    fun getAll(): List<Movie>
+}
